test(lib-sourcify): add zksolc/EraVM verification tests and reorganize
Add end-to-end verification tests for native zkSync EraVM (zksolc) contracts,
replaying 14 already-verified Abstract mainnet contracts against the public RPC
(ZkSolcVerification.e2e.spec.ts + committed source fixtures). Coverage spans the
keccak (<=1.5.12) and CBOR (>=1.5.13) metadata eras, legacy zksolc down to 1.3.8,
explicit and resolved era-solc editions, constructor args, and perfect/partial
runtime plus perfect/null creation matches. The suite self-skips when the RPC is
unreachable or ZKSYNC_E2E=false.

Also move the unit-level EraVM matching tests out of ZkSolcCompilation.spec.ts
into ZkSolcVerification.spec.ts, and extract the shared zksolc mock helpers into
test/utils/zksolcTestHelpers.ts. Add a real zksolc IZkSolcCompiler to test/utils.
