+++ packages/lib-sourcify/test/Verification/zksync-fixtures/ERC1967Proxy-8739d502/expected.json
@@ -1,0 +1,18 @@
+{
+  "chainId": 2741,
+  "address": "0x8739d5024b5143278e2b15bd9e7c26f6cec658f1",
+  "contractName": "ERC1967Proxy",
+  "contractPath": "@openzeppelin/contracts/proxy/ERC1967/ERC1967Proxy.sol",
+  "creationTransactionHash": "0x737f370e8d865a5d50ebb70a25364dd5b77e3bb46bc34aa4f782abd9a7dc6c5c",
+  "explorerZkSolcVersion": "v1.4.1",
+  "explorerSolcVersion": "v0.8.4+commit.c7e474f2",
+  "constructorArguments": "000000000000000000000000470d1c91b1b1d9295815a2357fb0d20e7350ab7100000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000204146faf77000000000000000000000000056f829183ec806a78c26c98961678c24fab71af0000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000018000000000000000000000000000000000000000000000000000000000000000015635979a221c34931e32620b9293a463065555ea71fe97cd6237ade875b12e9e0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003c000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000030000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000001a000000000000000000000000000000000000000000000000000000000000001a00000000000000000000000000000000000000000000000000000000000000036bb14509a612f01fbbc4cffeebd4bbfb492a86df717ebe92eb6df432a3f00a25f8cd23c2ab91237730770bbea08d61005cdda0984348f3f6eecb559638c0bba0e101faedac5851e32b9b23b5f9411a8c2bac4aae3ed4dd7b811dd1a72ea4aa7100000000000000000000000000000000000000000000000000000000",
+  "hasLibraries": false,
+  "directDeploy": true,
+  "creationMatchable": true,
+  "onchainRuntimeByteLength": 5280,
+  "compilerVersion": "zksolc:1.4.1;solc:v0.8.4+commit.c7e474f2",
+  "expectedRuntimeMatch": "perfect",
+  "expectedCreationMatch": "perfect",
+  "notes": "Legacy zksolc 1.4.1 with an upstream commit-form solc (v0.8.4+commit): exercises the pre-1.5.0 path with a plain (non era-solc) solc build. Direct ContractDeployer deploy with ABI-encoded constructor arguments; perfect runtime + perfect creation match."
+}
