+++ packages/lib-sourcify/test/Verification/zksync-fixtures/TransparentUpgradeableProxy-954ba822/expected.json
@@ -1,0 +1,18 @@
+{
+  "chainId": 2741,
+  "address": "0x954ba8223a6bfec1cc3867139243a02ba0bc66e4",
+  "contractName": "TransparentUpgradeableProxy",
+  "contractPath": "@openzeppelin/contracts/proxy/transparent/TransparentUpgradeableProxy.sol",
+  "creationTransactionHash": "0x891f1e505fa5c114d1ba50bd0d0cd5f9497bddcb95fd619e50923e89f552e3f0",
+  "explorerZkSolcVersion": "v1.3.18",
+  "explorerSolcVersion": "v0.8.20+commit.a1b79de6",
+  "constructorArguments": "0000000000000000000000000ba08b2e2fc141c4def0d80638486490dc50320500000000000000000000000039f71318a250d30aba4d04c6b55d6ba91b40135a00000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000084a31ee5b0000000000000000000000000d7f9f54194c633f36ccd5f3da84ad4a1c38cb2cb00000000000000000000000057891966931eb4bb6fb81430e6ce0a03aabde063010001211b0c33353cdf7a320f768e3dc40bce1326d639fcac099bba9ecd8e3400000000000000000000000039f71318a250d30aba4d04c6b55d6ba91b40135a00000000000000000000000000000000000000000000000000000000",
+  "hasLibraries": false,
+  "directDeploy": true,
+  "creationMatchable": true,
+  "onchainRuntimeByteLength": 11168,
+  "compilerVersion": "zksolc:1.3.18;solc:v0.8.20+commit.a1b79de6",
+  "expectedRuntimeMatch": "perfect",
+  "expectedCreationMatch": null,
+  "notes": "Legacy pre-1.5.0 zksolc (1.3.18): zksolc-bin fallback download + pre-1.5.0 path with an upstream commit-form solc (v0.8.20). Direct ContractDeployer deploy, but its EraVM-native creation tx (type 0xff) isn't retrievable through the standard eth_getTransaction RPC, so there is no creation match; perfect runtime."
+}
